trunk: fixing compilation warnings from recent g++ version
git-svn-id: https://svn.code.sf.net/p/kaldi/code/trunk@2716 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/feat/feature-fbank-test.cc
@@ -172,7 +172,7 @@ static void UnitTestHTKCompare1() {
   HtkHeader header = {
     kaldi_features.NumRows(),
     100000,  // 10ms
-    sizeof(float)*kaldi_features.NumCols(),
+    static_cast<int16>(sizeof(float)*kaldi_features.NumCols()),
     000007  // FBANK
   };
   {
